requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==35.0.0
Django==3.0.8
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
